requirements.txt
fastapi
uvicorn
pymongo
motor
pydantic
streamlit
python-dotenv